refactor: move physics/core modules to submodules and expose public api, unify i/o, extra tweaks
