+++ src/monarchs/core/debug.py
@@ -25,11 +25,11 @@ def neighbourhood_check(grid, target_row, target_col):
                 print(f"  Lake status: {cell['lake']}")
                 print(f"  Lid status: {cell['lid']}")
                 print(
-                    f"  Saturation status at surface = ",
+                    "  Saturation status at surface = ",
                     cell["saturation"][:5],
                 )
-                print(f"  Meltflag status at surface = ", cell["meltflag"][:5])
-                print(f"  Exposed water flag = ", cell["exposed_water"])
+                print("  Meltflag status at surface = ", cell["meltflag"][:5])
+                print("  Exposed water flag = ", cell["exposed_water"])
                 if not cell["valid_cell"]:
                     print("  WARNING: Cell is marked INVALID before physics starts.")
                 print("-------------------------")
