+++ src/monarchs/physics/lake/__init__.py
@@ -1,0 +1,31 @@
+"""
+Lake physics subpackage.
+
+Public API
+----------
+lake_formation           : Evolve shallow (<10 cm) exposed water toward a lake.
+lake_development         : Evolve a deep (>=10 cm) open lake.
+freeze_pre_lake          : Refreeze a shallow pre-lake into the firn column.
+radiative_transfer       : Compute solar radiation penetration into the lake.
+turbulent_mixing         : Compute turbulent heat exchange and boundary change.
+combine_lid_firn         : Merge lid + lake + firn into a single firn column.
+"""
+
+from monarchs.physics.lake.formation import lake_formation, freeze_pre_lake
+from monarchs.physics.lake.development import (
+    lake_development,
+    turbulent_mixing,
+    radiative_transfer,
+)
+from monarchs.physics.lake.reset_column import combine_lid_firn
+from monarchs.physics.lake import reset_column
+
+__all__ = [
+    "lake_formation",
+    "lake_development",
+    "freeze_pre_lake",
+    "turbulent_mixing",
+    "radiative_transfer",
+    "combine_lid_firn",
+    "reset_column",
+]
