+++ src/monarchs/physics/timestep.py
@@ -6,16 +6,10 @@ accordingly.
 """
 
 import numpy as np
-from monarchs.physics import (
-    firn_column,
-    lake,
-    solver,
-    lid,
-    percolation,
-    virtual_lid,
-    snow_accumulation,
-    reset_column,
-)
+from monarchs.physics.firn import firn_column, percolation, snow_accumulation
+from monarchs.physics import lake, solver
+from monarchs.physics.lid import lid, virtual_lid
+from monarchs.physics.lake import reset_column
 from monarchs.physics.constants import rho_ice, rho_water
 from monarchs.core.error_handling import generic_error, check_correct
 
@@ -172,7 +166,6 @@ def timestep_loop(cell, dt, met_data, t_steps_per_day, toggle_dict):
                     lake.lake_formation(cell, dt, met_data[t_step])
 
             elif cell["lake"] and not cell["lid"]:
-
                 if lake_development_toggle:
                     Fu = lake.lake_development(cell, dt, met_data[t_step])
                 # TODO - add refreezing calculation here - relevant if we have
@@ -197,7 +190,6 @@ def timestep_loop(cell, dt, met_data, t_steps_per_day, toggle_dict):
                         reset_column.combine_lid_firn(cell, freeze_lake=False)
 
             elif cell["lake"] and cell["lid"]:
-
                 if lid_development_toggle:
                     # turn virtual lid off if full lid present
                     cell["v_lid"] = False
@@ -220,8 +212,8 @@ def timestep_loop(cell, dt, met_data, t_steps_per_day, toggle_dict):
                 #              melting for half a day, and therefore has
                 #              significant slush/water on the surface)
                 if cell["lake_depth"] <= 1e-5:
-                    #print("Combining lid and firn column")
-                    #print("Lid melt count:", cell["lid_melt_count"])
+                    # print("Combining lid and firn column")
+                    # print("Lid melt count:", cell["lid_melt_count"])
                     reset_column.combine_lid_firn(cell, surface_slush=False)
                 elif cell["lid_melt_count"] > 12:
                     reset_column.combine_lid_firn(cell, surface_slush=True)
@@ -231,7 +223,7 @@ def timestep_loop(cell, dt, met_data, t_steps_per_day, toggle_dict):
         # region of liquid water inside the firn. We need to ensure that this
         # continues to freeze based on the conditions what we take to be our
         # new "firn column" - much as it did when it was a lake under a lid.
-        # Effectively, we need to mimick the behaviour of the lid refreezing,
+        # Effectively, we need to mimic the behaviour of the lid refreezing,
         # so that the water doesn't just sit there forever.
 
         # If we have Sfrac + Lfrac > 1, we need to ensure that Lfrac is
@@ -257,8 +249,7 @@ def timestep_loop(cell, dt, met_data, t_steps_per_day, toggle_dict):
                 and cell["Sfrac"][0] <= 1
             ):
                 message = (
-                    "Lfrac + Sfrac > 1 after regridding, after saturation"
-                    " calculation."
+                    "Lfrac + Sfrac > 1 after regridding, after saturation calculation."
                 )
                 generic_error(cell, routine_name, message)
         if not ignore_errors:
